feat: refactor CitizenInvite component to replace SpouseSubmit with SubmitDynastyModal and remove unused UnderEighteenMember component
